Add confirmation before deleting notes

+++ src/redux/hooks/noteHooks.ts
@@ -69,6 +69,9 @@ export function useNoteRemove() {
       if (user === null) {
         return;
       }
+      if (!confirm('This operation cannot be undone')) {
+        return;
+      }
 
       dispatch(noteRemove.request(noteId));
